+++ control-plane/api-gateway/binding/validation.go
@@ -3,6 +3,7 @@ package binding
 import (
 	"strings"
 
+	"github.com/hashicorp/consul/api"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	klabels "k8s.io/apimachinery/pkg/labels"
@@ -12,7 +13,8 @@ import (
 	gwv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 )
 
-func (r *routeBinder[T, U]) validateRefs(namespace string, refs []gwv1beta1.BackendRef) routeValidationResults {
+// validateRefs
+func validateRefs(namespace string, refs []gwv1beta1.BackendRef, services map[types.NamespacedName]api.CatalogService) routeValidationResults {
 	var result routeValidationResults
 	for _, ref := range refs {
 		nsn := types.NamespacedName{
@@ -32,7 +34,7 @@ func (r *routeBinder[T, U]) validateRefs(namespace string, refs []gwv1beta1.Back
 			continue
 		}
 
-		if _, found := r.services[nsn]; !found {
+		if _, found := services[nsn]; !found {
 			result = append(result, routeValidationResult{
 				namespace: nsn.Namespace,
 				backend:   ref,
